+++ konfuzio_sdk/urls.py
@@ -1,4 +1,4 @@
-"""Endpoints of the Konfuzio Host."""
+"""URLs to the endpoints of Konfuzio Host."""
 
 import logging
 
@@ -6,6 +6,9 @@ from konfuzio_sdk import KONFUZIO_HOST, KONFUZIO_PROJECT_ID
 from typing import Union
 
 logger = logging.getLogger(__name__)
+
+
+# TOKEN-AUTH
 
 
 def get_auth_token_url(host: str) -> str:
@@ -17,34 +20,71 @@ def get_auth_token_url(host: str) -> str:
     return f"{host}/api/token-auth/"
 
 
-def get_project_list_url(host: str = KONFUZIO_HOST) -> str:
+# PROJECTS
+
+
+def get_projects_list_url(host: str = KONFUZIO_HOST) -> str:
     """
-    Generate URL to load all the projects available for the user.
+    Generate URL to list all the projects available for the user.
 
     :return: URL to get all the projects for the user.
     """
     return f"{host}/api/projects/"
 
 
-def create_new_project_url(host: str = KONFUZIO_HOST) -> str:
+def get_project_url(project_id: Union[int, None] = None, host: str = KONFUZIO_HOST) -> str:
     """
-    Generate URL to create a new project.
+    Generate URL to access the project details.
 
+    :param project_id: ID of the project
     :param host: Konfuzio host
-    :return: URL to create a new project.
+    :return: URL to access the project details.
     """
-    return f"{host}/api/projects/"
+    project_id = project_id if project_id else KONFUZIO_PROJECT_ID
+    return f'{host}/api/projects/{project_id}/'
 
 
 def get_documents_meta_url(host: str = KONFUZIO_HOST, project_id: int = KONFUZIO_PROJECT_ID) -> str:
     """
-    Generate URL to load meta information about documents.
+    Generate URL to load meta information about the documents in the project.
 
     :param host: Konfuzio host
     :param project_id: ID of the project
     :return: URL to get all the documents details.
     """
     return f"{host}/api/projects/{project_id}/docs/"
+
+
+def get_document_annotations_url(
+    document_id: int, host: str = KONFUZIO_HOST, project_id: int = KONFUZIO_PROJECT_ID
+) -> str:
+    """
+    Access annotations of a document.
+
+    :param document_id: ID of the document as integer
+    :param host: Konfuzio host
+    :param project_id: ID of the project
+    :return: URL to access the annotations of a document
+    """
+    return f'{host}/api/projects/{project_id}/docs/{document_id}/annotations/'
+
+
+def get_document_segmentation_details_url(
+    document_id: int, project_id: int = KONFUZIO_PROJECT_ID, host: str = KONFUZIO_HOST, action='segmentation'
+) -> str:
+    """
+    Generate URL to get the segmentation results of a document.
+
+    :param document_id: ID of the document as integer
+    :param project_id: ID of the project
+    :param host: Konfuzio host
+    :param action: Action from where to get the results
+    :return: URL to access the segmentation results of a document
+    """
+    return f'{host}/api/projects/{project_id}/docs/{document_id}/{action}/'
+
+
+# DOCUMENTS
 
 
 def get_upload_document_url(host: str = KONFUZIO_HOST) -> str:
@@ -57,30 +97,20 @@ def get_upload_document_url(host: str = KONFUZIO_HOST) -> str:
     return f"{host}/api/v2/docs/"
 
 
-def update_document_url(document_id: int, host: str = KONFUZIO_HOST) -> str:
+def get_document_url(document_id: int, host: str = KONFUZIO_HOST) -> str:
     """
-    Generate URL to update a document.
+    Generate URL to access a document.
 
     :param document_id: ID of the document as integer
     :param host: Konfuzio host
-    :return: URL to update a document
+    :return: URL to access a document
     """
     return f"{host}/api/v2/docs/{document_id}/"
 
 
-def get_create_label_url(host: str = KONFUZIO_HOST) -> str:
-    """
-    Generate URL to create a label.
-
-    :param host: Konfuzio host
-    :return: URL to create a label.
-    """
-    return f"{host}/api/v2/labels/"
-
-
 def get_document_ocr_file_url(document_id: int, host: str = KONFUZIO_HOST) -> str:
     """
-    Generate URL to access OCR version of document.
+    Generate URL to get the OCR version of the document.
 
     :param document_id: ID of the document as integer
     :param host: Konfuzio host
@@ -91,7 +121,7 @@ def get_document_ocr_file_url(document_id: int, host: str = KONFUZIO_HOST) -> st
 
 def get_document_original_file_url(document_id: int, host: str = KONFUZIO_HOST) -> str:
     """
-    Generate URL to access original version of the document.
+    Generate URL to get the original version of the document.
 
     :param document_id: ID of the document as integer
     :param host: Konfuzio host
@@ -108,7 +138,7 @@ def get_document_api_details_url(
     extra_fields='bbox',
 ) -> str:
     """
-    Generate URL to access document details of one document in a project.
+    Generate URL to access the details of a document in a project.
 
     :param document_id: ID of the document as integer
     :param host: Konfuzio host
@@ -123,68 +153,43 @@ def get_document_api_details_url(
     )
 
 
-def get_project_url(project_id: Union[int, None] = None, host: str = KONFUZIO_HOST) -> str:
-    """
-    Generate URL to get project details.
+# LABELS
 
-    :param project_id: ID of the project
+
+def get_labels_url(host: str = KONFUZIO_HOST) -> str:
+    """
+    Generate URL to list all labels for the user.
+
     :param host: Konfuzio host
-    :return: URL to get project details.
+    :return: URL to list all labels for the user.
     """
-    project_id = project_id if project_id else KONFUZIO_PROJECT_ID
-    return f'{host}/api/projects/{project_id}/'
+    return f"{host}/api/v2/labels/"
 
 
-def post_project_api_document_annotations_url(
-    document_id: int, host: str = KONFUZIO_HOST, project_id: int = KONFUZIO_PROJECT_ID
-) -> str:
+def get_label_url(label_id: int, host: str = KONFUZIO_HOST) -> str:
     """
-    Add new annotations to a document.
+    Generate URL to access a label.
 
-    :param document_id: ID of the document as integer
+    :param label_id: ID of the label as integer
     :param host: Konfuzio host
-    :param project_id: ID of the project
-    :return: URL for adding annotations to a document
+    :return: URL to access a label
     """
-    return f'{host}/api/projects/{project_id}/docs/{document_id}/annotations/'
+    return f"{host}/api/v2/labels/{label_id}/"
 
 
-def delete_project_api_document_annotations_url(
+# ANNOTATIONS
+
+
+def get_annotation_url(
     document_id: int, annotation_id: int, host: str = KONFUZIO_HOST, project_id: int = KONFUZIO_PROJECT_ID
 ) -> str:
     """
-    Delete the annotation of a document.
+    Generate URL to access an annotation.
 
     :param document_id: ID of the document as integer
     :param annotation_id: ID of the annotation as integer
     :param host: Konfuzio host
     :param project_id: ID of the project
-    :return: URL to delete annotation of a document
+    :return: URL to access an annotation of a document
     """
-    return f'{host}/api/projects/{project_id}/docs/{document_id}/' f'annotations/{annotation_id}/'
-
-
-def get_document_result_v1(document_id: int, host: str = KONFUZIO_HOST) -> str:
-    """
-    Generate URL to access web interface for labeling of this project.
-
-    :param document_id: ID of the document as integer
-    :param host: Konfuzio host
-    :return: URL for labeling of the project.
-    """
-    return f'{host}/api/v1/docs/{document_id}/'
-
-
-def get_document_segmentation_details_url(
-    document_id: int, project_id: int = KONFUZIO_PROJECT_ID, host: str = KONFUZIO_HOST, action='segmentation'
-) -> str:
-    """
-    Generate URL to get the segmentation results of a document.
-
-    :param document_id: ID of the document as integer
-    :param project_id: ID of the project
-    :param host: Konfuzio host
-    :param action: Action from where to get the results
-    :return: URL to access the segmentation results of a document
-    """
-    return f'{host}/api/projects/{project_id}/docs/{document_id}/{action}/'
+    return f'{host}/api/projects/{project_id}/docs/{document_id}/annotations/{annotation_id}/'